Fix imports following refactor
