+++ test/public/qcFlags/detailsForDataPass.test.js
@@ -27,7 +27,7 @@ const {
     waitForTableLength,
 } = require('../defaults.js');
 const { resetDatabaseContent } = require('../../utilities/resetDatabaseContent.js');
-const { navigateToRunsPerDataPass } = require('../runs/navigationUtils.js');
+const { navigateToRunsPerDataPass } = require('../runsPerPeriodOrPass/dataPassesUtilities.js');
 
 const { expect } = chai;
