+++ test/public/qcFlags/forDataPassCreation.test.js
@@ -28,7 +28,7 @@ const {
     getTableContent,
 } = require('../defaults.js');
 const { resetDatabaseContent } = require('../../utilities/resetDatabaseContent.js');
-const { navigateToRunsPerDataPass } = require('../runs/navigationUtils.js');
+const { navigateToRunsPerDataPass } = require('../runsPerPeriodOrPass/dataPassesUtilities.js');
 
 const { expect } = chai;
